Provide utilities to allow dpi dependent zooming of fonts
This contribution extends the support for dynamical adaption on SWT applications on DPI related zoom changes in the win32 implementation by adding support to provider and cache scaled variants of Font. It consists of:
- internal font registries to enable reusage of scaled variants of fonts. If a font is disposed from external, it will be automatically recreated on demand. There are two registries:
- a default registry to mimic the existing behavior, that all fonts are scaled to the native zoom of the primary monitor
- a registry to create and manage scaled variants of fonts
The SWTFontProvider is configured based on the "swt.autoScale.updateOnRuntime"-flag to use one of the two font registries. The creation of system fonts is adapted to use the provider

Contributes to https://github.com/eclipse-platform/eclipse.platform.swt/issues/62
and https://github.com/eclipse-platform/eclipse.platform.swt/issues/127